+++ packages/fuels-contract/src/execution_script.rs
@@ -1,6 +1,8 @@
 use crate::contract_calls_utils::{
-    extract_message_outputs, extract_unique_contract_ids, extract_variable_outputs,
-    get_single_call_instructions, CallOpcodeParamsOffset,
+    convert_to_signed_resources, extract_message_outputs, extract_unique_asset_ids,
+    extract_unique_contract_ids, extract_variable_outputs, generate_asset_change_outputs,
+    generate_contract_inputs, generate_contract_outputs, get_single_call_instructions,
+    sum_up_amounts_for_each_asset_id, CallOpcodeParamsOffset,
 };
 use anyhow::Result;
 use std::collections::HashSet;
@@ -13,19 +15,23 @@ use fuels_core::{offsets::call_script_data_offset, parameters::TxParameters};
 use fuels_signers::provider::Provider;
 use fuels_signers::{Signer, WalletUnlocked};
 
-use fuel_tx::field::Inputs;
+use fuel_tx::field::{Inputs, Outputs};
 use fuels_core::tx::ContractId;
 use fuels_types::errors::Error;
+use itertools::chain;
 use std::vec;
 
 use crate::contract::ContractCall;
 use crate::contract_calls_utils::{
     build_script_data_from_contract_calls, calculate_required_asset_amounts, get_instructions,
-    prepare_script_inputs_outputs,
 };
 
 /// [`TransactionExecution`] provides methods to create and call/simulate a transaction that carries
-/// out contract method calls or script calls
+/// out contract method calls or script calls.The [`ExecutableFuelCall`] structure contains a
+/// [`pub tx: fuels_core::tx::Script`] field that can be modified.
+/// Using this structure is recommended only when the user has a thorough understanding of what he
+/// is doing, since incorrect settings during execution may result in errors
+///
 #[derive(Debug)]
 pub struct ExecutableFuelCall {
     pub tx: fuels_core::tx::Script,
@@ -59,9 +65,9 @@ impl PrepareFuelCall {
         }
     }
 
-    /// Creates a [`TransactionExecution`] from contract calls. The internal [`Transaction`] is
+    /// Creates a [`PrepareFuelCall`] from contract calls. The internal [`Transaction`] is
     /// initialized with the actual script instructions, script data needed to perform the call and
-    /// transaction inputs/outputs consisting of assets and contracts.
+    /// transaction inputs/outputs consisting of assets and contracts
     pub async fn from_contract_calls(
         calls: &[ContractCall],
         tx_parameters: &TxParameters,
@@ -102,11 +108,6 @@ impl PrepareFuelCall {
         Ok(PrepareFuelCall::new(tx, calls, wallet.clone()))
     }
 
-    pub async fn prepare(&mut self) -> Result<ExecutableFuelCall, Error> {
-        prepare_script_inputs_outputs(self).await?;
-        Ok(ExecutableFuelCall::new(self.tx.clone()))
-    }
-
     pub fn add_inputs(&mut self, inputs: Vec<Input>) -> &mut PrepareFuelCall {
         self.inputs.extend(inputs);
         self
@@ -115,6 +116,87 @@ impl PrepareFuelCall {
     pub fn add_outputs(&mut self, outputs: Vec<Output>) -> &mut PrepareFuelCall {
         self.outputs.extend(outputs);
         self
+    }
+
+    /// This function creates an [`ExecutableFuelCall`] from a [`PrepareFuelCall`], which avoids
+    /// the possibility of duplication of inputs and outputs.
+    /// For this reason, we strongly recommend the use of this function
+    pub async fn prepare(&mut self) -> Result<ExecutableFuelCall, Error> {
+        Self::prepare_inputs_outputs(self).await?;
+        Ok(ExecutableFuelCall::new(self.tx.clone()))
+    }
+
+    pub async fn prepare_inputs_outputs(script: &mut PrepareFuelCall) -> Result<(), Error> {
+        let mut spendable_resources = vec![];
+
+        let script_inputs = script
+            .inputs
+            .iter()
+            .filter(|input| !matches!(input, Input::Contract { .. }))
+            .map(|values| (*values.asset_id().unwrap(), values.amount().unwrap()))
+            .collect::<Vec<_>>();
+
+        let (script_message_output, script_coin_output): (Vec<_>, Vec<_>) = script
+            .outputs
+            .iter()
+            .cloned()
+            .partition(|input| matches!(input, Output::Message { .. }));
+
+        let merged_inputs = sum_up_amounts_for_each_asset_id(
+            chain!(script.calls.required_asset_amounts.clone(), script_inputs).collect::<Vec<_>>(),
+        );
+
+        for (asset_id, amount) in &merged_inputs {
+            let resources = script
+                .wallet
+                .get_spendable_resources(*asset_id, *amount)
+                .await?;
+            spendable_resources.extend(resources);
+        }
+
+        let asset_ids = extract_unique_asset_ids(&spendable_resources);
+        *script.tx.inputs_mut() = chain!(
+            generate_contract_inputs(script.calls.calls_contract_ids.clone()),
+            convert_to_signed_resources(spendable_resources),
+        )
+        .collect::<Vec<Input>>();
+
+        // Note the contract_outputs need to come first since the
+        // contract_inputs are referencing them via `output_index`. The node
+        // will, upon receiving our request, use `output_index` to index the
+        // `inputs` array we've sent over.
+        *script.tx.outputs_mut() = chain!(
+            generate_contract_outputs(script.calls.calls_contract_ids.len()),
+            generate_asset_change_outputs(script.wallet.address(), asset_ids),
+            chain!(
+                script.calls.calls_variable_outputs.clone(),
+                script_coin_output
+            ),
+            chain!(
+                script.calls.calls_message_outputs.clone(),
+                script_message_output
+            ),
+        )
+        .collect::<Vec<Output>>();
+
+        let base_asset_amount = merged_inputs
+            .iter()
+            .find(|(asset_id, _)| *asset_id == AssetId::default())
+            .map(|(_, amount)| *amount)
+            .unwrap_or(0);
+
+        script
+            .wallet
+            .add_fee_resources(&mut script.tx, base_asset_amount, 0)
+            .await?;
+
+        script
+            .wallet
+            .sign_transaction(&mut script.tx)
+            .await
+            .unwrap();
+
+        Ok(())
     }
 }
 
@@ -149,7 +231,7 @@ impl ExecutableFuelCall {
             .iter()
             .any(|r|
                 matches!(r, Receipt::ScriptResult { result, .. } if *result != ScriptExecutionResult::Success)
-        ) {
+            ) {
             return Err(Error::RevertTransactionError(Default::default(), receipts));
         }
